qmk_gui: Fix crash on flashing on Chinese language system
Would crash in to_str with
'utf-8' code can't decode byte ...

That's because "VolumeName" the volume name would include Chinese
characters. They should be decodeable by utf8, so not sure why it
happens.
But we don't actually use this in the output.

Also note that the columns aren't orderd the way we ask them, but
alphabetically.

Signed-off-by: Daniel Schaefer <dhs@frame.work>